-added SystemTray API -added initial mDNS API -added regex support -added UUID support -removed old FindRadonFramework.cmake file -removed unnecessary includes in CMakeLists.txt -replaced function typedef by new using syntax -fixed concurrency algorithm -replaced private constructor/destructor by new delete syntax -fixed some Compiler macros added a noexcept macro -fixed a bug in string literal constructor of String class if the string is larger as the local buffer -Atomic variable allow now to add a specified value -use override syntax on virtual functions now -concurrent algorithm are working with AutoPointerArray now -added additional util functions to IPAddress class -refactored Server class -reintroduced socket options -added realpath support -added swap memory function -added open with default application function -added string formated time function -introduced Text and Util sub-namespace to RadonFramework -fixed UnitTest, ResourceTest and SystemTest

+++ include/RadonFramework/System/Threading/Interlocked.hpp
@@ -51,7 +51,7 @@ struct Interlocked
     /// Performs a atomic addition operation.
     static RF_Type::Int64 InterlockedAdd64(
         RF_Type::Int64 volatile* Addend,
-        RF_Type::Int32 Value);
+        RF_Type::Int64 Value);
 
     // Logical operations
     /// Performs an atomic AND operation on the specified Int8 values.
